requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
click==8.1.7
lxml==5.3.0